feat(authorization): add user authorization

+++ src/graphql/User.ts
@@ -38,34 +38,6 @@ export const UsersQuery = extendType({
   },
 });
 
-export const CreateUserQuery = extendType({
-  type: "Mutation",
-  definition(t) {
-    t.nonNull.field("createUser", {
-      type: "User",
-      args: {
-        name: nonNull(stringArg()),
-        email: nonNull(stringArg()),
-        password: nonNull(stringArg()),
-      },
-      resolve(_, { name, email, password }, { prisma }) {
-        try {
-          return prisma.user.create({
-            data: {
-              id: Date.now().toString(),
-              name,
-              email,
-              password,
-            },
-          });
-        } catch (err) {
-          throw err;
-        }
-      },
-    });
-  },
-});
-
 export const UserQuery = extendType({
   type: "Query",
   definition(t) {
@@ -93,20 +65,24 @@ export const UpdateUser = extendType({
     t.nonNull.field("updateUser", {
       type: "User",
       args: {
-        id: nonNull(idArg()),
         name: stringArg(),
         email: stringArg(),
       },
-      resolve: async (_, { id, name, email }, { prisma }) => {
-        const match = await prisma.user.findUnique({ where: { id } });
+      resolve: async (_, { name, email }, { prisma, userId }) => {
+        if (!userId) {
+          throw new Error("not logged in");
+        }
+        const match = await prisma.user.findUnique({
+          where: { id: userId.toString() },
+        });
         if (!match) {
-          throw new Error(`user by id: ${id} does not exist`);
+          throw new Error(`user does not exist`);
         }
 
         return prisma.user.update({
-          where: { id },
+          where: { id: userId.toString() },
           data: {
-            id,
+            id: userId.toString(),
             name: name || match?.name,
             email: email || match?.email,
           },
@@ -121,19 +97,20 @@ export const DeleteUser = extendType({
   definition(t) {
     t.nonNull.field("deleteUser", {
       type: "User",
-      args: {
-        id: nonNull(idArg()),
-      },
 
-      resolve: async (_, { id }, { prisma }) => {
-        const links = await prisma.link.deleteMany({
+      resolve: async (_, __, { prisma, userId }) => {
+        if (!userId) {
+          throw new Error("not logged in");
+        }
+
+        await prisma.link.deleteMany({
           where: {
-            postedById: id,
+            postedById: userId.toString(),
           },
         });
 
         const user = await prisma.user.delete({
-          where: { id },
+          where: { id: userId.toString() },
         });
 
         return user;
